+++ pkg/controller/standard.oam.dev/v1alpha1/metrics/metricstrait_controller_integration_test.go
@@ -18,7 +18,7 @@ import (
 	"k8s.io/utils/pointer"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 
-	"github.com/oam-dev/kubevela/apis/v1alpha1"
+	"github.com/oam-dev/kubevela/apis/standard.oam.dev/v1alpha1"
 )
 
 var (
