+++ pkg/controller/standard.oam.dev/v1alpha1/podspecworkload/suite_test.go
@@ -30,7 +30,7 @@ import (
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
-	v1alpha1 "github.com/oam-dev/kubevela/apis/v1alpha1"
+	v1alpha1 "github.com/oam-dev/kubevela/apis/standard.oam.dev/v1alpha1"
 	// +kubebuilder:scaffold:imports
 )
 
@@ -55,7 +55,7 @@ var _ = BeforeSuite(func(done Done) {
 	By("bootstrapping test environment")
 	testEnv = &envtest.Environment{
 		CRDDirectoryPaths: []string{
-			filepath.Join("../../../..", "charts/vela-core/crds"), // this has all the required CRDs,
+			filepath.Join("../../../../..", "charts/vela-core/crds"), // this has all the required CRDs,
 			filepath.Join("..", "config", "crd", "bases")},
 	}
