+++ pkg/controller/standard.oam.dev/v1alpha1/routes/ingress/contour_ingress_test.go
@@ -12,7 +12,7 @@ import (
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/utils/pointer"
 
-	standardv1alpha1 "github.com/oam-dev/kubevela/apis/v1alpha1"
+	standardv1alpha1 "github.com/oam-dev/kubevela/apis/standard.oam.dev/v1alpha1"
 )
 
 func TestContourConstruct(t *testing.T) {
