+++ pkg/controller/standard.oam.dev/v1alpha1/routes/suite_test.go
@@ -40,8 +40,8 @@ import (
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
-	standardv1alpha1 "github.com/oam-dev/kubevela/apis/v1alpha1"
-	"github.com/oam-dev/kubevela/pkg/controller/v1alpha1/podspecworkload"
+	standardv1alpha1 "github.com/oam-dev/kubevela/apis/standard.oam.dev/v1alpha1"
+	"github.com/oam-dev/kubevela/pkg/controller/standard.oam.dev/v1alpha1/podspecworkload"
 	// +kubebuilder:scaffold:imports
 )
 
@@ -75,7 +75,7 @@ var _ = BeforeSuite(func(done Done) {
 	useExistCluster := false
 	testEnv = &envtest.Environment{
 		CRDDirectoryPaths: []string{
-			filepath.Join("../../../..", "charts/vela-core/crds"), // this has all the required CRDs,
+			filepath.Join("../../../../..", "charts/vela-core/crds"), // this has all the required CRDs,
 		},
 		UseExistingCluster: &useExistCluster,
 	}
